move components into a common folder
