+++ src/components/common/index.js
@@ -1,0 +1,9 @@
+import Text from './Text'
+import Link from './Link'
+import Dropdown from './Dropdown'
+
+export {
+  Text,
+  Link,
+  Dropdown,
+}
