--- src/components/index.js
@@ -1,9 +1,0 @@
-import Text from './text'
-import Link from './link'
-import Dropdown from './dropdown'
-
-export {
-  Text,
-  Link,
-  Dropdown,
-}
